requirements.txt
openai==0.27.8
HTMLParser>=0.0.2
PyQRCode>=1.2.1
qrcode>=7.4.2
requests>=2.28.2
chardet>=5.1.0
Pillow
pre-commit
web.py
filetype
arrow>=1.2.3
openpyxl>=3.1.2
croniter>=1.4.1
wework
python-magic
python-magic-bin
